agents/birdwatcher/requirements.txt
agno>=1.5.9
google-cloud-aiplatform>=1.38.0
anthropic>=0.7.0
requests>=2.31.0
urllib3>=1.26.0
httpx>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
mcp>=1.9.4
rich>=13.0.0
typing-extensions>=4.5.0
pydantic>=2.0.0
resend
slack-sdk>=3.20.0
cryptography
python-dotenv>=1.0.0
slack-bolt>=1.23.0
aiohttp>=3.9.0
ipdb>=0.13.13